refactor: 네이밍 Drawing2 -> Drawing
