+++ app/src/main/java/woowacourse/paint/drawing/Rectangle.kt
@@ -3,10 +3,10 @@ package woowacourse.paint.drawing
 import android.graphics.Canvas
 import android.graphics.Paint
 import android.graphics.RectF
-import woowacourse.paint.drawing.Drawing2.Companion.DEFAULT_BRUSH_CAP
-import woowacourse.paint.drawing.Drawing2.Companion.DEFAULT_BRUSH_STYLE
+import woowacourse.paint.drawing.Drawing.Companion.DEFAULT_BRUSH_CAP
+import woowacourse.paint.drawing.Drawing.Companion.DEFAULT_BRUSH_STYLE
 
-data class Rectangle(val rect: RectF = RectF(), private val paint: Paint) : Drawing2 {
+data class Rectangle(val rect: RectF = RectF(), private val paint: Paint) : Drawing {
     override fun setUpDefaultPaint() {
         paint.apply {
             style = DEFAULT_BRUSH_STYLE
@@ -39,7 +39,7 @@ data class Rectangle(val rect: RectF = RectF(), private val paint: Paint) : Draw
         }
     }
 
-    override fun copyWithPaint(thickness: Float): Drawing2 {
+    override fun copyWithPaint(thickness: Float): Drawing {
         val paint =
             Paint(paint).apply {
                 strokeWidth = thickness
@@ -47,7 +47,7 @@ data class Rectangle(val rect: RectF = RectF(), private val paint: Paint) : Draw
         return Rectangle(rect, paint)
     }
 
-    override fun copyWithPaint(color: Int): Drawing2 {
+    override fun copyWithPaint(color: Int): Drawing {
         val paint =
             Paint(paint).apply {
                 this.color = color
@@ -58,7 +58,7 @@ data class Rectangle(val rect: RectF = RectF(), private val paint: Paint) : Draw
     override fun copyPoint(
         pointX: Float,
         pointY: Float,
-    ): Drawing2 = this.copy(rect = RectF(pointX, pointY, pointX, pointY))
+    ): Drawing = this.copy(rect = RectF(pointX, pointY, pointX, pointY))
 
     companion object {
         fun default(): Rectangle {
